Use Gazelle to manage go dependencies
Following the instructions from
https://github.com/bazelbuild/bazel-gazelle#update-repos

Why is this a good change:

I, for one, really dislike having to manually find the go package sum
for my dependencies. This change means that you can use the normal
golang tools to add a 3rd party dependency. Simply add the dependency as
an import in a file in the repo, run `go get ...` inside the directory
that has the new import (which adds the dependency to go.mod) and then
update your `deps.bzl` file by running

```
bazelisk run //:gazelle -- \
  update-repos \
  -from_file=go.mod \
  -to_macro=deps.bzl%install_buildbuddy_dependencies
```

this further improves the harmony with the golang tooling, and for me
allows code completion tools to autocomplete function names, arguments,
and even goto definition.

N.B. any time you modify go.mod you will need to run the command that I
left commented out at the top of `deps.bzl` to have it reflect the
change.