Expand environment variables in config
## What?
Adds expansion of environment variables in the BuildBuddy config file.

## Why?
My main motivation for this to be able to keep secrets out of the configmap and inject them via environment variables.

## How?
Using the `os.ExpandEnv` function which expands all environment variables that are of the form `${ENV_VAR}` in a string.