Add configurable inmemory cache for API Key -> Group information.
Bazel can generate bursts of RPCs when doing builds and we need to
perform auth for each one. We currently hit the database for every
request even though this information rarely changes.

This change caches lookup results for a fixed, configurable amount of
time which is simple but good enough to handle the RPC bursts.