Fix FileReader for disk util
The returned Reader offered no facility for closing the file after
reading was complete; this has been corrected. Additionally, the
LimitReader was not being passed the length and was instead being given
the entire file size when a length was specified; this has been changed
to use a SectionReader with the specified length. The change to a
sectionReader also affords us Seek and ReadAt functionality with no
further effort, so the returned interface includes those capabilities.