Fix shallow copy of proto
Protos are not meant to be copied in Go (see
https://stackoverflow.com/questions/64183794/why-do-the-go-generated-protobuf-files-contain-mutex-locks).
Here we instead Merge the proto to be copied into an empty proto. It's
possibly correct to use proto.Clone instead, but I couldn't figure out
how to coerce the return value from the generic Message interface to the
specific proto type without using reflection.