Refactor auth to separate authentication & context creation. (#538)
Refactor auth to separate authentication & context creation.

I would like to re-authorize executors periodically to verify that
the API key is still valid. AuthenticateGRPCRequest is renamed to
AuthenticatedGRPCContext and the new AuthenticateGRPCRequest method
always performs a new lookup based on the credentials present in the
request w/o looking at JWT present in the context.

Part of addressing buildbuddy-io/buildbuddy-internal#496

The issue with the previous attempt was that the user metadata ("user
token") key was not being set in the context which would lead to user
creation to fail as it relies on this key to be present.