Add prober that runs Bazel over a generated workspace.
The generated workspace includes rules that generate actions that copy
their inputs to their outputs.

This PR also adds a new container image that contains the cloudprober
binary, the bazelrbe prober binary as well as the bazel binaries.

Lastly this also changes console logging to write to stderr instead of
stdout which is a common convention.
(cloudprober consumes prober stdout for any custom metrics that the
prober wants to provide.)