Add some basic task reservation balancing across groups.
This PR creates a separate priority queue per group ID and then does a
round robin over the PQs to dequeue the next task.

This PR also adds a Group ID label to the remote_execution_queue_length
metric. This should not impact scaling as the metric is already
aggregated by other fields when being exposed to k8s.