Add support for multiple "partitions" in disk cache.
The disk cache config has a list of partition configs consisting
of an ID and a maximum size. The ID is used to determine where on disk
the cache partition is rooted & the max size is used to instantiate the
LRU.

The config also has a list of mappings which specify when a partition
should be used. Each mapping consists of a Group ID and an remote
instance name prefix. Partition will be used if a cache request is
received with matching group ID and a remote instance name that has a
matching prefix.

Finally, a `WithRemoteInstanceName` method is introduced to allow the
instance name to be passed to the cache implementation so it can make a
decision about which partition to use.