Add WithIsolation method to the Cache interface.
WithIsolation is an instruction to the cache implementation to return an
accessor for the cache that isolates the data for the given cache type +
remote instance name combination from any other combination.

This method will supersede WithPrefix, which can be removed once the
distributed disk implementation has been migrated to the new method.