Add support & migration tooling for a "v2" disk cache layout. (#915)
* Add support & migration tooling for a "v2" disk cache layout.

Under the v1 layout, files are structured as follows:
/disk_cache/GR1234/abababababab
/disk_cache/GR1234/prefix/abababababab
/disk_cache/PTFOO/GR4567/abababababab
/disk_cache/PTFOO/GR4567/prefix/abababababab

This layout has the following issues:
 - All digests for the same prefix are stored in a single directory
which can bump up against ext4 limits for sufficiently large number of
digests.
 - Digests in the "default" partition are stored directly under the root
unlikely the other partitions which have their own subdirectories. This
requires special handling to ignore the non-default partitions under the
root.

Under the v2 layout, the same data is structured as follows:
/disk_cache/PTdefault/GR1234/abab/abababababab
/disk_cache/PTdefault/GR1234/prefix/abab/abababababab
/disk_cache/PTFOO/GR4567/abab/abababababab
/disk_cache/PTFOO/GR4567/prefix/abab/abababababab

Digests are stored under a subdirectory consisting of the first 4
characters of the hash.

Default partition digests are stored under an explicit PTdefault root
consistent with other partitions.