Host side firecracker changes to support CASFS.
We start a ByteStream server proxy on the host that proxies requests
from the guest to the remote cache using the current execution context
and its credentials. A future PR will take advantage of the file cache.

I made SendExecRequestToGuest public so we can use it from the `vmstart`
tool to create VM + setup CASFS.