On executor shutdown re-enqueue tasks that are still in the executor queue.  (#1101)
Upon receiving a shutdown signal, the executor will send a message to the scheduler informing it about the tasks that are still in its queue. The scheduler will then re-enqueue those tasks to other executors.

This is a shorter-term solution that we can ship with tomorrow's release. It didn't look like the more comprehensive changes would make the release tomorrow.

Note that during rollout we may see some harmless Invalid message from executor warnings if an old app receives a shut down request it's not expecting.

Part of addressing buildbuddy-io/buildbuddy-internal#900