Only add unclaimed tasks on the "master" scheduler.
Prior to this change, the app receiving the Execute RPC would add the
unclaimed task to the set but so would other schedulers that receive an
enqueue RPC from the master scheduler. This would cause a race between
the task claim and the enqueues. e.g.

   scheduler1 adds task
   scheduler1 sends enqueue RPCs to 3 other schedulers
   the task gets claimed concurrently with one of the enqueue RPCs
   if the task claim wins the race, the enqueue RPC will re-add the task
      back to the set

I also added a mechanism to automatically trim any stale tasks.