Don't publish COMPLETED events when tasks are retried. (#1159)
* Don't publish COMPLETED events when tasks are retried.

We are currently sending a COMPLETED event to Bazel as soon as the first
attempt fails even if we re-schedule the task to be retried.

`ExecuteTaskAndStreamResults` does not publish COMPLETED events if the
error should be retried. It also returns the retry flag so it can be
propagated to the TaskLeaser which uses it to determine whether the task
should be re-enqueued.

When retries are exhausted in `SchedulerServer`, it calls
`MarkExecutionFailed` on the `ExecutionServer` to tell it to mark the
execution as failed and send the update to Bazel.

`ReEnqueueTaskRequest` gains a 'reason' field so that the the last error
can be propagated to the user.

I had to made some changes to the leasing protocol to support these
changes. The executor can now request that a lease be 'released' instead
of being 'finalized'. This releases the claim on the task without
deleting the task and prevents the scheduler from re-enqueing the task.
The previous behavior caused retries to be enqueued by both the
scheduler and the executor since the executor did not gracefully close
the lease stream in case of retries.

Part of addressing
https://github.com/buildbuddy-io/buildbuddy-internal/issues/855
https://github.com/buildbuddy-io/buildbuddy-internal/issues/939