Switch executor pool metrics to use the 'generic' metrics.
We currently query the metrics that are explicitily defined per pool.

Moving to the generic metrics will fix discrepancies in the prod data due
to counting the custom customer executors in the default pool graphs.

After this roll out we can also remove the custom pool scrape config
from Prometheus.