Refactor SDKROOT calculation
We now store a map of sdk -> sdkPath (e.g. `iPhoneSimulator15.2` -> `Platforms/iPhoneSimulator.platform/Developer/SDKs/iPhoneSimulator15.2.sdk`, `MacOSX.sdk` -> `SDKs/MacOSX.sdk`). We now also return an error if we can't find the desired SDK in the desired Xcode version, which is more correct than falling back to a version-less SDK.

This change is more performant, with a map lookup instead of iterating over an array. But more importantly, it is needed for a future change to support using the Xcode Command Line Tools as the default "Xcode" version.