Make Xcode matching more percise
If the following Xcode versions are installed:

- 12.0.0.12A7209
- 12.1.0.12A7403
- 12.1.1.12A7605b
- 12.2.0.12B45b

Then map these input versions to these output versions:

- 12 - > 12.2.0.12B45b
- 12.0 -> 12.0.0.12A7209
- 12.0.0 -> 12.0.0.12A7209
- 12.0.0.12A7209 -> 12.0.0.12A7209
- 12.1 -> 12.1.1.12A7605b
- 12.1.0 -> 12.1.0.12A7403
- 12.1.0.12A7403 -> 12.1.0.12A7403
- 12.1.1 -> 12.1.1.12A7605b
- 12.1.1.12A7605b -> 12.1.1.12A7605b
- 12.2 -> 12.2.0.12B45b
- 12.2.0 -> 12.2.0.12B45b
- 12.2.0.12B45b -> 12.2.0.12B45b

To summarize, we map a less percise version to the highest available version it could match. This is good.

We _also_ currently take a fully qualified version, such as  12.5.0.12E262 and if it doesn't match one of the inputs available above, we make it less percise until it potentially matches: 12.5.0, 12.5, and finally 12. This means that 12.5.0.12E262 -> 12 - > 12.2.0.12B45b, which is invalid.

This commit removes that final step, meaning in the example above, 12.5.0.12E262 will return an error saying that Xcode version isn't available.