Add ability to configure Redis Ring (sharded) client instead of single server-client. (#1589)
* Add ability to configure Redis Ring (sharded) client instead of single-server client.

The Ring client does not support all the same options that the simple
client does which means we can't create a ring client everywhere.
Notably we support unix sockets in redis target which are not supported
by the ring client.

The go-redis library also does not have a common option type that works
for both the simple and ring clients. It has a a UniversalOptions type
but it works for everything EXCEPT the Ring client so we introduce our
own in redisutil.