Add option to mount workspace ext4 fs instead of extracting contents. (#1627)
Mounting the filesystem file directly is significantly faster.

This is behind a flag for now because it doesn't work well locally due
to needing root. I tried using `udisksctl` to get around this
requirement but ran into the following issues:
   1) On local Ubuntu, at least on my
      installation, something automatically mounts the filesystem as soon as I
      set up the loop device even before I call "mount".
   2) On a deployed image, we would need to have DBus & udisk daemons
      installed and running for this to work.

I'm going to mess around with user namespaces to see if we can get the
same behavior locally as we do when running in a container on GKE. That
might be a better path forward.