Redis unavailability handling. (#1592)
* Redis unavailability handling.

This PR adds the ability to detect if execution related data has disappeared
because Redis restarted and lost data or because a Redis shard went
away. The detection is built into the PubSub subscription. The PubSub
subscriber monitors the state of the PubSub stream in the background
and publishes an error to the subscription client if the PubSub stream
disappears.

If we detect that a PubSub stream has gone away, we generate a synthetic
operation failure message to Bazel and rely on Bazel to retry execution
of the action.