Add package for starting test executor binary.
The new `testexecutor` package is similiar to the existing `app` package
for starting the app binary.

The common behavior has been moved to the `testserver` package.

I also moved the `app.FreePort` function into a new `testport` package
since it doesn't logically belong to the app package and I want to use
it from the testexecutor package.