Explicitly concatenate slice flags and set them
Some slice flags do not have a way to distinguish their astring
representation of an empty slice and a slice containing one zero value
(looking at you, stringSliceFlag). To avoid any errors this could cause,
set the values explicitly to the concatenated slice instead of depending
on `Set` to convert and concatenate a string representation of the flag.