Fix log flicker when chunk is finalized.
When a chunk is finalized, the volatile portion of the log moves from
the current chunk to the next chunk. If we draw the finalized chunk
without delay, the volatile portion is removed and then printed again
when the next (live) chunk retrieved. To avoid this flicker, we delay
printing the finalized chunk until we retrieve the next chunk with the
volatile log.