Cancel RBE actions when invocation is interrupted.
When a BES interrupted event is received, we ask the execution server to
cancel any pending executions for that invocation.

The execution server uses the data in the database to retrieve the list
of actions that are currently executing.

The execution server uses this list to ask the scheduler server to
cancel these tasks.

The scheduler server cancels the tasks by deleting them from Redis. The
task lease mechanism is updated to verify that the task still exists
when the lease is renewed.