Runner side changes to extract run target metadata.
If the `--record_run_metadata` flag is passed to the runner, we perform
the bazel run with the `--script_path` argument which does all the
normal run preparation followed by writing a run script without actually
running it.

We then parse the generated script to extract metadata necessary for the
CLI to be able to run the binary:
 - the default arguments
 - walk the runtime files and make sure they are in the CAS

This information is then recorded in the build event stream using the
new RunTargetAnalyzed event.