Pass through extra args to `bazel run` to the downloaded binary.
If the user executes `bazel run //target -- --foo=bar`, we need to pass
`--foo=bar` to the downloaded binar.

When parsing the command line args, we split them into regular &
passthrough slices (stored in `BazelArgs` struct).

This also fixes a missing `defer` in the ci_runner (introduced in the
last PR).