findDefaultDevice -> findDevice (#1980)
Add a flag to config the device to find. This can be 'default' or the
internal ip range of the subnet, e.g. 172.24.0.0/18.

Tested: set the flag to 172.24.0.0/18; and push the executor to my own
test executor pool which already have two networks configured: default
and executor with subnet 172.24.0.0/18. Run a bazel test that pings an
internal ip address in the default network and ping www.google.com.
Results:
  - Test fails when running on podman b/c it can successfully ping the
  internal ip address (10.138.0.88)
  https://app.buildbuddy.dev/invocation/acef757d-30e9-4989-a9b0-0d6804565d72.
  - Test fails when running on firecracker in dev executor pool.
  (firecracker is on default network) It can
  successfully ping the internal ip address.
  https://app.buildbuddy.dev/invocation/5aa5ee5d-7de3-4430-856b-f8ba769da6db
  - Test succeeds when running on firecracker (with this change) in dev
  executor lulu pool. (firecracker is on the new executor network) It
  fails to ping the internal ip address.
  https://app.buildbuddy.dev/invocation/70712bdb-e14d-4a39-bd6a-ce5019c9ac89