Merge concurrent identical action execution requests. (#2016)
* Merge concurrent identical action execution requests.

If we have already scheduled action digest foo and get another request
to execute action foo, then return the ID of the existing execution
instead of scheduling a new execution.

When we receive an execution request, we first check whether there's a
{group_id, instance_name, action_digest} -> execution_id mapping
available in Redis. If there is one and the execution_id is valid
according to the scheduler, then we return the existing execution ID to
the client.

For new executions, we insert a {group_id, instance_name, action_id} ->
execution_id mapping into Redis as well as a reverse mapping of
execution_id -> {group_id, instance_name, action_id}. This reverse
mapping is necessary because when we update executions we don't always
know the correct group_id to use.

One limitation to be aware of is that the execution will only be
visible in the UI for the first invocation that scheduled the execution
because within our current data model an execution ID can only be
associated with a single invocation. We can address this with a separate
change.

https://github.com/buildbuddy-io/buildbuddy-internal/issues/1352