Support associating executions with more than one invocation.
This PR introduces a new table, InvocationExecutions, that stores
Invocation -> Execution associations.

For backwards compatibility, when querying executions we query both
using the new table as well any entries in the executions table that do
not have a link in the new table.

After some time passes, we can stop querying for 'unlinked' executions
since it's unlikely that remote execution details would be important for
old invocations.