Podman changes to support image streaming.
On executor startup, we start the stargz store as a child process and
configured podman to look for image information in the store.

When a new pull is requested, we consult the registry to obtain the
optimized image name. This pull is also used for revalidating the
container credentials.

This change support registry auth, but the CAS data is still not
segregated by group. That will require another change.