Utility to dedupe expensive work across servers. (#2147)
* Utility to dedupe expensive work across servers.

Every client for a given work key will try to obtain a lock in Redis.
The first client to get the lock will perform the work and publish the
result to a stream. If the owner does not refresh the lock within 5
seconds, another worker will have a chance to claim the lock and do the
work.