add schema for quota configs (#2117)
* add schema for quota configs

QuotaBucket is to define the sustained rate and max burst.
QuotaGroup is to define the relationship between group ID / IP and the
QuotaBucket.

Note:
To save the space, for each namespace, we will define a default bucket. If a group id / ip is not in the QuotaGroup table, then quota manager will apply the default bucket.