Increase the number of fallback peers we attempt to use for handoff. (#2400)
Prior to this change, the number of fallback peers we attempted to use
was equal to the number of failed preferred peers. In the most common
case of 1 failed peer, we would only try one fallback peer. If that peer
was also down, we would give up and this would typically lead to a
failed write.

This change returns up to 3 fallback peers before giving up.