Redact secrets for flagz (#2601)
<!--
Optional: Provide additional description (beyond the PR title).
Description should provide any background or motivation needed for the change, as well
as a high-level overview of the approach taken (if the change is not straightforward).
Detailed rationale for specific sections of the code are probably better off as code comments
so that future readers of that code can benefit.
-->

Leverages `autoflags.New` to tag flags as being secret so that they can be redacted in flagz.

Refactors `autoflags` so that deprecation and marking secret can easily be used in concert using the tags.

Refactors `flagutil/yaml` so that we can remove entries from documented YAML and so that we can use information from the flag in `DocumentNodeOption`s.

These refactors allow us to implement a new option, `RedactSecrets`, which uses hints from struct tags and the flag, if present, to determine if a field needs to be redacted and remove it if it does.

---

**Version bump**: Minor <!-- Required. Choose from: Major, Minor, Patch, None -->

<!-- See https://semver.org/#semantic-versioning-specification-semver. Summary:
* Major: Breaking change that causes existing functionality to not work as expected.
* Minor: Non-breaking change that adds functionality (examples: new feature; new API options)
* Patch: Non-breaking change that fixes an issue, improves performance, or refactors
         code.
* None:  Changed files are not included in releases (tests, docs, development setup,
         production configs)
-->

<!-- Optional:
**Related issues**: Fixes #1, Unblocks #2 ...
-->