Add a Shutdown function for build event handler
Create a child context with Cancel function when we open a channel; and
put the cancel function in a map in build event handler. When the server
receives shutdown signal, we will call cancel functions.

In BuildEventServer, when the context from the channel is done, we will
return UnavailableError to signal bazel to retry.