Add AWS IAM support for DB connections.
To support this functionality, users will need to use the new 'advanced'
configuration stanzas. We need this to be able to generate the DSN using
the short-lived tokens returned by AWS.

A typical use will look like this:

```
 advanced_data_source:
    driver: mysql
    use_aws_iam: true
    region: some-region
    endpoint: some-instance:3306
    username: some-user
    db_name: buildbuddy
```

Most users can continue to use the simpler 'data_source' option.

The other notable change is that we now control the creation of new DB
connections so that we can generate a new DSN every time since the
generated AWS IAM tokens are short-lived.