Improve split performance by cloning the underlying pebble DB.
Before splitting, we start a new cluster that will become the new right
side. This is unchanged from before, but because the pebble DB is
already open on the new cluster we have to swap the DB with the cloned
data after the replica is already running.

After the clone operation is done, we force dragonboat library to take a
Raft snapshot and to compact the Raft log. This "tricks" the library
into using the replica-provided snapshot data for recovery operations.
This is important because the real Raft log does not contain the keys
that were cloned.