vmexec: increase sleep duration memory test
When this test action is executed in an Executor that is under
a heavy load, the polling mechanism we use to monitor resource
consumption of the vm could be delayed and thus, reporting inaccurate
results.

This is hard to reproduce as our executors autoscale on kubernetes.
This is also low critical as we are currently not using this metric for
any billing related functionality.

De-flake the test for now by increasing the sleep duration.
Also mark the test as flaky to notify future maintainer of this issue.