bazel: fix //enterprise package visibility
We organize `//enterprise/...` code to have a different license compare to
code in `//... -//enterprise/...`. Yet, there are packages outside of
`enterprise` dir that depends on the code inside `enterprise` directory.

This PR is the first step in correcting that problem. Ensure that all
code within `//enterprise/...` has a strict visibility set to
`//enterprise:__subpackages__`. Exceptions are strictly stated so that
we could query for them in the future and apply appropriate refactoring.

Before this PR
```
> bazel query 'visible(//..., //enterprise/...)' | wc -l
29
```

After this PR
```
> bazel query 'visible(//..., //enterprise/...)' | wc -l
4
```