blobstore: support path_prefix for storing blobs
This is a follow up from #3597 where were added a feature to allow users
to specify a path prefix in their S3 to store cache objects. This
enables users to configure BuildBuddy with a single bucket.

Apply the same change to blobstore so user could have this configuration
for build events as well.

Closes #3595