tools/rbeperf: move under //enterprise/tools
This represent an alternative approach to #3671

RBE is an enterprise-only feature of BuildBuddy. Our bench test tool
rbeperf has dependencies on //enterprise source code and therefore,
could not be built from FOSS version of the repo.

Because our internal repo reference this tool, there will be a short
period of time where CI in our internal repo is broken. A separate PR
will be sent to update the references in our internal repo.