Incorporate encryption support into pebble cache.
We store a `bool` property on the group to indicate whether a customer
has encryption enabled. This property is looked up and cached as part of
the normal auth check so that we can avoid doing any additional database
lookups to determine if auth is enabled.

Cache partitions have an additional property to indicate whether
encrypted content can be stored on that partition. We allow partitions
(e.g. the default one) to share CAS data across customers which is
incompatible with encryption. We will require dedicated customer
partitions before encryption can be enabled.