runner_test: fix timeout on darwin
After investigation in https://github.com/buildbuddy-io/buildbuddy/pull/3701,
it's shown that on average, Shutdown_RunnersReturnRetriableOrNilError
could only finish 40+ trials out of 100 on Intel Mac. This is after we
have sharded the runner_test action to the exact test count (18) so that
each test func is run on it's own shard with dedicated CPU.

The root cause of the slowness is with how the test is constructed and
run. Our test would attempt to spin up 50 execution tasks asynchronously
in separate goroutines. Then in the main goroutune, wait until the task
count hit 50/2=25 and shutdown the runner pool arruptedly to check for
error. Repeat this trial 100 times and ensure that in total it took less
than 60 seconds to complete.

This works well if the test execution is provided with a generous
amount of CPU to manage 50+ goroutines, but on our CI, the test is given
a limited amount of CPU to work with. On Darwin CI, the test is often
executed in parallel with other tests, giving it less resources to
manage multiple goroutines. This caused each iteration(trial) to take more
than 1 second to complete, making the cost of 100 trials to be beyond the
60 seconds timeout.

Upon tagging the test with `exclusive`, the shard running this test
would finish 100 trials in 44 seconds.

Reduce the trial count to 30 to avoid timeout.