rules_go: simplified Go SDK config and provide a go binary alias
Simplified our Go SDK config and provide some context doc on why we need
to set it up the way it is. Upgrading Go version should be relatively
simpler after this.

Also provide an alias to the "go" binary in the go_sdk so we don't need
to remember what platform we are on while calling it.
It's useful to run `go fmt` or `go mod` commands using the exact Go
version Bazel use.

Co-authored-by: Brandon Duffany <brandon@buildbuddy.io>