Use ANSI quotes for queries and quote table names (#3821)
<!-- Optional: Provide additional context (beyond the PR title). -->

The motivation for this is that postgres does not and cannot use backticks to
indicate identifiers such as table names, so if we want to support postgres, we
need to switch to double quotes for that purpose (sql_mode="ANSI_QUOTES" in
mysql). We also need to quote table names, since we use mixed-case table names,
postgres is case sensitive, and unquoted identifiers are all converted to lower
case by postgres. This also means all strings must be specified with single
qoutes (`'`).

<!-- Optional: link a GitHub issue.
     Example: "Fixes #123" will auto-close #123 when the PR is merged. -->

**Related issues**: N/A