Include encryption key in pebble key.
Turning encryption on or off should be a seamless operation from Bazel's
perspective. If we don't include this information in the key, we'd have
to get the information from the value which would negatively impact
certain APIs (e.g. FindMissing) both in performance and complexity.

For example, if the user turns off encryption we want FindMissing to not
report previously written encrypted blobs as present which would require
unmarshalling and checking the value for each key.

We have to adjust the eviction sampling to take into account the fact
that there could be more than one key with the same digest prefix. Note
that this can also be true in the new AC key format if the same digest
appears under multiple remote instance name hashes.