scheduler_server: flip remove_stale_executors to true
Not setting this flag to true would make BuildBuddy iterate through
the stale executors during scheduling.  With sufficient amount of
stale executors, scheduling process could be significantly slower.

Since we have set this flag to true in production for a while, let's
deprecate it by setting the default to true in up-coming release and
remove the flag entirely in a subsequent release.