Add new pebble key format that includes the encryption key ID.
This format adds an optional encryption key segment for keys that refer
to encrypted contents.

This version also makes the "remote instance hash" segment always be
present even if the remote instance name is empty. Instead of omitting
the segment, we use a "0" value. This simplifies the parsing logic as
the encryption key is also optional.

Adding the encryption key ID to the key makes it simple to deal with
transitions from having encryption enabled to off, or even if the
encryption is turned off and on again.

For example, if a user has encryption enabled and then turns it off, we
want previously written encrypted artifacts to be treated as
non-existent. Without having the encryption metadata in the key, we
would have to unmarshal and check every value during FindMissing calls
which would affect the performance and complexity of the implementation.
Adding the encryption key ID to the key simplifies this since the keys
would be different when encryption is off and when on (or even when
turned off and re-enabled with a different key).