Update sampling code to account for keys with the same digest prefix.
With the new AC format and with encryption, it becomes possible for
multiple keys to have the same digest in their prefix. Our previous
sampling implementation was built with the assumption that there's only
one key for a single digest.

Here we update the sampling algorithm to iterate sequentially from the
first key until we reach a key where the hash is different so that we
include all pebble keys with the same hash in the eviction pool.