Fix esbuild pointing to non-existent directory
When attempting to upgrade rules_nodejs for the internal repo, it seems
that the location of `node_modules` directory could be moved to be under
the Open Source repo (this repo).

Attempting to detect the usual 'node_modules' directory not being there
and switch to use the new path under 'external/'.

This fix is expected to be a no-op in the current repo and the current
internal repo.  However this should help unblock this internal PR:
https://github.com/buildbuddy-io/buildbuddy-internal/pull/2270