Make group ID sampling work for ANON data.
Keys for ANON data don't follow the normal key format so they won't be
found when sampling keys with the "GR" prefix.

To make the process work for ANON keys, we make up a random pseudo group
ID for anon keys that's only used when sampling group IDs. When
sampling, we check if we would have hit this group ID had it really
existed in the key space and if so return the ANON group as the sampled
group id.