Add cdc chunking support to PebbleCache.
When chunking is enabled, during write, we will first attempt to
chunk it, and then compress each individual chunk, and finally encrypt the chunk
if encryption is enabled.  If the data is compressed, we will first decompress it
before we chunk it.

If there is only one single chunk, we will write the chunk inline,
similar to how we write small files inline when chunking is disabled.
If there are multiple chunks, we will write a file record in Pebble. which include
all the resource names that points to the chunks. Every chunks will be
stored inline in Pebble.

Added unit tests.