config: Propagate parse failures upward
This change returns the parse error upward when a config fails to parse,
instead of early-returning no error. This fixes an issue where a
malformed config results in a confusing failure to launch:

```
2023/07/07 22:32:10.053 INF BuildBuddy v2.12.49 (933606adb4ed2c807eab7a9edcc27782eb0e1d10) compiled with go1.20.5 X:nocoverageredesign
2023/07/07 22:32:10.054 INF Reading buildbuddy config from '/local/config.yaml'
2023/07/07 22:32:10.062 INF Auto-migrating DB
2023/07/07 22:32:10.078 INF Successfully configured sqlite database.
2023/07/07 22:32:10.079 WRN No authentication will be configured: rpc error: code = Internal desc = OIDC authenticator failed to configure: rpc error: code = FailedPrecondition desc = No auth providers specified in config!
2023/07/07 22:32:10.080 FTL rpc error: code = Internal desc = Invalid Remote Execution Redis config.
```

This is happening becuase the parse failure essentially causes the
config to be discarded, at which point default values are used, which
happens to result in an invalid configuration.

With this change, Buildbuddy will still fail to launch, but will print
the actual config parse error instead.

Tested: Built an image with this change and used to debug a config that
wouldn't launch properly after updating (spoiler: it was malformed the
whole time)