bytebufferpool: store pointer to slice to reduce allocation
sync.Pool should be storing pointers instead of struct.
A Go slice is a struct underneath (with one of the fields being the
pointer to an array).

Swap out sync.Pool usage in bytebufferpool with a `safeSyncPool` struct
that provides type-safe access to pool's values.
Change bytebufferpool.Pool to use this generic wrapper with type being
pointer to byte slice.