runner: allow setting podID using environment variable
When cgroup v2 or other container orchestration is used, it's possible
that podID is not set in `/proc/*/cpuset` and become empty.

Let's provide a way for user to explicitly set podID from K8s container
definition via environment variable.

For example, one could use the following config

```yaml

  containers:
    - name: buildbuddy-executor
      ...
      env:
      - name: K8S_POD_ID
        valueFrom:
          fieldRef:
            fieldPath: metadata.uid
```

to set this value.