Add a CDC Chunker.
Add a Chunker that implements io.WriteCloser.

When we create a Chunker, we passes two write functions to indicate how
to write the data when there is a single chunk versus when there are
multiple chunks. For example, when there is a single chunk, we want to
write it inline directly, but when there are multiple chunks, we want to
write a file level record that includes pointers to individual chunks.

Note: The fastcdc-go library is not thread safe because when
fastcdc.NewChunker is called it will change the global variable table.

So I included a patch that don't change the table when the seed is not
set (or equals zero). This should be functionally the same, since
x ^ 0 = x.