docker: use ContainerAttach to capture std{out,err}
Our previous approach was prone to race conditions between the container
starting up and our call to `ContainerLogs` API. In the case where we call
the log API before container actually started, we get no outputs.  This
resulted in flaky tests observed in our CI.

Switch to use `ContainerAttach` API to capture the container outputs as
they are streamed out to reliably obtain the logs.